feat(microsalt): Added analysis microsalt command
- Moved link_microbial from analysis.py to microbial.py
